chore: move event into client/event
其为 client 运行时产生的结构, 理应放置其下
