+++ client/event/friend.go
