+++ client/event/group.go
