+++ client/event/self.go
